pyproject.toml
[tool.poetry]
name = "taverncrier"
version = "1.0.0"
description = ""
authors = ["Justin <justin@aperturebot.science>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
betterdisco-py = {git = "https://github.com/elderlabs/BetterDisco.git", rev = "staging/justin", extras = ["performance", "yaml"]}
psycopg2-binary = "^2.9.9"
peewee = "^3.17.1"
pyyaml = "^6.0.1"
redis = "^5.0.4"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
